# Fix StringIndexOutOfBoundsException issue in JDBC test framework